Fix simulation bug and pretify outputs

+++ src/main/scala/vinted/Constants.scala
@@ -2,10 +2,10 @@ package vinted
 
 object Constants:
 
-  type KG_CO2 = Int
-  type KM = Int
-  type EURO = Int
-  type NB_ITEMS_BY_MONTH = Int
+  type KG_CO2 = Double
+  type KM = Double
+  type EURO = Double
+  type NB_ITEMS_BY_MONTH = Double
   type MONTH = Int
 
   val KG_CO2_PER_ITEM_PRODUCTION = 15.0
